Move FractionalVotingPower to its own mod
