+++ shared/src/types/vote_extensions/ethereum_events.rs
@@ -2,11 +2,8 @@
 //! in vote extensions.
 
 use std::collections::{HashMap, HashSet};
-use std::ops::{Add, AddAssign};
 
 use borsh::{BorshDeserialize, BorshSchema, BorshSerialize};
-use eyre::{eyre, Result};
-use num_rational::Ratio;
 
 use crate::proto::Signed;
 use crate::types::address::Address;
@@ -47,96 +44,6 @@ impl EthEventsVext {
     /// and return the signed data.
     pub fn sign(self, signing_key: &common::SecretKey) -> Signed<Self> {
         Signed::new(signing_key, self)
-    }
-}
-
-/// A fraction of the total voting power. This should always be a reduced
-/// fraction that is between zero and one inclusive.
-#[derive(Clone, PartialOrd, Ord, PartialEq, Eq, Debug)]
-pub struct FractionalVotingPower(Ratio<u64>);
-
-impl FractionalVotingPower {
-    /// Two thirds of the voting power.
-    pub const TWO_THIRDS: FractionalVotingPower =
-        FractionalVotingPower(Ratio::new_raw(2, 3));
-
-    /// Create a new FractionalVotingPower. It must be between zero and one
-    /// inclusive.
-    pub fn new(numer: u64, denom: u64) -> Result<Self> {
-        if denom == 0 {
-            return Err(eyre!("denominator can't be zero"));
-        }
-        let ratio: Ratio<u64> = (numer, denom).into();
-        if ratio > 1.into() {
-            return Err(eyre!(
-                "fractional voting power cannot be greater than one"
-            ));
-        }
-        Ok(Self(ratio))
-    }
-}
-
-impl Default for FractionalVotingPower {
-    fn default() -> Self {
-        Self::new(0, 1).unwrap()
-    }
-}
-
-impl From<&FractionalVotingPower> for (u64, u64) {
-    fn from(ratio: &FractionalVotingPower) -> Self {
-        (ratio.0.numer().to_owned(), ratio.0.denom().to_owned())
-    }
-}
-
-impl Add<FractionalVotingPower> for FractionalVotingPower {
-    type Output = Self;
-
-    fn add(self, rhs: FractionalVotingPower) -> Self::Output {
-        Self(self.0 + rhs.0)
-    }
-}
-
-impl AddAssign<FractionalVotingPower> for FractionalVotingPower {
-    fn add_assign(&mut self, rhs: FractionalVotingPower) {
-        *self = Self(self.0 + rhs.0)
-    }
-}
-
-impl BorshSerialize for FractionalVotingPower {
-    fn serialize<W: ark_serialize::Write>(
-        &self,
-        writer: &mut W,
-    ) -> std::io::Result<()> {
-        let (numer, denom): (u64, u64) = self.into();
-        (numer, denom).serialize(writer)
-    }
-}
-
-impl BorshDeserialize for FractionalVotingPower {
-    fn deserialize(buf: &mut &[u8]) -> std::io::Result<Self> {
-        let (numer, denom): (u64, u64) = BorshDeserialize::deserialize(buf)?;
-        Ok(FractionalVotingPower(Ratio::<u64>::new(numer, denom)))
-    }
-}
-
-impl BorshSchema for FractionalVotingPower {
-    fn add_definitions_recursively(
-        definitions: &mut std::collections::HashMap<
-            borsh::schema::Declaration,
-            borsh::schema::Definition,
-        >,
-    ) {
-        let fields =
-            borsh::schema::Fields::UnnamedFields(borsh::maybestd::vec![
-                u64::declaration(),
-                u64::declaration()
-            ]);
-        let definition = borsh::schema::Definition::Struct { fields };
-        Self::add_definition(Self::declaration(), definition, definitions);
-    }
-
-    fn declaration() -> borsh::schema::Declaration {
-        "FractionalVotingPower".into()
     }
 }
 
@@ -227,37 +134,6 @@ mod tests {
                 80, 38, 224, 23
             ])
         );
-    }
-
-    /// This test is ultimately just exercising the underlying
-    /// library we use for fractions, we want to make sure
-    /// operators work as expected with our FractionalVotingPower
-    /// type itself
-    #[test]
-    fn test_fractional_voting_power_ord_eq() {
-        assert!(
-            FractionalVotingPower::TWO_THIRDS
-                > FractionalVotingPower::new(1, 4).unwrap()
-        );
-        assert!(
-            FractionalVotingPower::new(1, 3).unwrap()
-                > FractionalVotingPower::new(1, 4).unwrap()
-        );
-        assert!(
-            FractionalVotingPower::new(1, 3).unwrap()
-                == FractionalVotingPower::new(2, 6).unwrap()
-        );
-    }
-
-    /// Test error handling on the FractionalVotingPower type
-    #[test]
-    fn test_fractional_voting_power_valid_fractions() {
-        assert!(FractionalVotingPower::new(0, 0).is_err());
-        assert!(FractionalVotingPower::new(1, 0).is_err());
-        assert!(FractionalVotingPower::new(0, 1).is_ok());
-        assert!(FractionalVotingPower::new(1, 1).is_ok());
-        assert!(FractionalVotingPower::new(1, 2).is_ok());
-        assert!(FractionalVotingPower::new(3, 2).is_err());
     }
 
     /// Test decompression of a set of Ethereum events
